removed legacy v_int parsing code from GScanner, use v_int64 solely now.
Thu Feb 26 20:37:57 2004  Tim Janik  <timj@gtk.org>

        * glib-extra.[hc]: removed legacy v_int parsing code from GScanner,
        use v_int64 solely now. renamed creation function to g_scanner_new64(),
        to catch old glib scanner uses.

        * testsfi.c:
        * sfistore.c:
        * sficomport.c: 64bit integer scanning fixes.

        * sfidl-parser.[hc]*: 64bit parsing ans Constant handling fixes, allow '+'
        as number sign besides '-'.

svn path=/trunk/; revision=2047

+++ sfi/sfidl-parser.hh
@@ -96,7 +96,7 @@ struct Constant {
 
   std::string str;
   float f;
-  int i;
+  gint64 i;
 };
 
 struct Param {
